Rename the table type module
