+++ wasm/datatypes/__init__.py
@@ -4,7 +4,7 @@ from .limits import (  # noqa: F401
 from .memory_type import (  # noqa: F401
     MemoryType,
 )
-from .tabletype import (  # noqa: F401
+from .table_type import (  # noqa: F401
     FuncRef,
     TableType,
 )
