+++ wasm/datatypes/table_type.py
